fix getscom_abs and _getscom code
Instead of reading the scom registers, and storing the results
to a memory address specified by o_data, the two routines were
loading data into a vector register, without storing it to the
o_data pointer, basically losing it.

There is no memory to memory PPC instruction, so the minimum number
of instructions to execute the code is two: a vector load and a vector store.

- Had to modify the ppe42_scom.h to make the _m_data parameter to the
PPE_LVD and PPE_LVDX macros an output only. Otherwise, the compiler
initializes the temp variable, adding two (useless) instructions.

- make the PPE_LVDX macro have the same flavor as PPE_LVD by eliminating
the pointer to m_data from both the _getscom function side and the
macro implementation side.

This fix is already merged in the hw/ppe repository.

Change-Id: I6390d2c5e2fbcecf440080917611a82ae88d313b
RTC: 135989
RTC: 137704
RTC: 136459
Reviewed-on: http://gfw160.aus.stglabs.ibm.com:8080/gerrit/20976
Reviewed-by: William A. Bryan <wilbryan@us.ibm.com>
Tested-by: William A. Bryan <wilbryan@us.ibm.com>
Reviewed-by: Fadi Kassem <fmkassem@us.ibm.com>