Fix occ/ssx_checkpoint_panic_and_save_ffdc bugs and create two deathbeds
1. use r3 to index the lmw instruction. r0 is not a valid index register
for this command ((RA|0) in the 405 user's manual). This maintains
the integrity of the r1 register, and hence the correct restoration of
all registers.

2. store the LR to the stack at the entrance of called functions,
and restoring fron the stack before calling the blr return instruction.

3. to avoid infinite loop through the occ/ssx_checkpoint_panic_and_save_ffdc
routines, two branch to self (deathbed) locations were created. In multichip
systems, this will avoid competing for the FFDC scom registers, which may be
causing an FFDC SCOM access timeout due to SCOM_PROTOCOL_ERROR_GETSCOM_BUSY
error (0x00726615).

Change-Id: Ifc1616212b4abdbce83008c4ba1ea16289ddaab2
CQ: SW324506
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/21101
Tested-by: FSP CI Jenkins
Reviewed-by: William A. Bryan <wilbryan@us.ibm.com>