Fix Sensor Location bug
AMEC_SENSOR_LOC_UNKNOWN is invalid sensor location, since it is defined as 0.
If an additional location is needed to define any driven/virtual sensor, a
new entry should be added to the AMEC_SENSOR_LOC enum (e.g., AMEC_SENSOR_LOC_FW),
and that new location have a 1 in its bit vector. This way, the existing
amester code would be ble to handle it properly in the initial sensor scan
routine.

Change-Id: I61dbd77fcadff2f81acfc42fde682f3659654b93
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/38664
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Martha Broyles <mbroyles@us.ibm.com>
Reviewed-by: William A. Bryan <wilbryan@us.ibm.com>