imageHdrScript: Fix binary I/O on integers of incorrect size
The variables were declared as "unsigned long", but the reads and
writes expected them to be of size 32 bits, so on a 64-bit system the
sizes did not match.  This could be observed on a 64-bit big endian
host where only zeroes were written to the files.

I also found and fixed some size mismatches in sscanf.