Move PWRP/M fields to start of EXTN block in poll response
On BMC systems, the kernel does not add additional sysfs files as
additional fields are added in subsequent poll responses. This can lead
to the PWRP and PWRM fields not being found by occ-control and then the
data will not be updated on DBUS.

This change will move the PWRP and PWRM fields to the beginning of the
EXTN block to prevent them from being missed.

Change-Id: Ifafbfe9c3c693f770ca4c2a9d5249fa24b2a71b0
STGD: 748657
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/c/occ/+/196720
Reviewed-by: Martha Broyles <mbroyles@us.ibm.com>
Reviewed-by: Chris Cain <cjcain@us.ibm.com>
Tested-by: Chris Cain <cjcain@us.ibm.com>